refactor: moved css files in assets
