+++ src/App.js
@@ -1,5 +1,5 @@
 //import logo from './logo.svg';
-import './App.css';
+import './assets/css/App.css';
 import Header from './components/Header'
 import Home from './pages/Home'
 import Footer from './components/Footer';
